+++ changelog.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## 2023-01-15
 
+### Changed
+- disable csrf token check
+
 ### Added
 - add unique constraint to fields in Machine and Snack model
 - add signal to auto set status of Machine
@@ -15,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - create management commands
     - create_sample: to create simple sample data
     - easy_setup: auto migrate and create simple sample data
+- add api endpoint for Snack and Machine Model
 
 
 ## 2023-01-14